Use new plugin architechure in listener
